used float64 instead of float32 to improve precision

+++ model.py
@@ -30,11 +30,11 @@ class VolterraModel(object):
         self.real_output = None
 
     def _init_data_structs(self):
-        self.input = tf.placeholder(tf.float32, [self.batch_size, self.model_memory], name="input")
+        self.input = tf.placeholder(tf.float64, [self.batch_size, self.model_memory], name="input")
         for order in range(1, self.model_order + 1):
             # init kernels, first order M, second M^2, ... M^N, where M is model memory and N is model order
             dimension = self.model_memory ** order
-            self.kernels[order] = tf.get_variable("kernel_order_{}".format(order), shape=[dimension], dtype=tf.float32, initializer=tf.random_normal_initializer())
+            self.kernels[order] = tf.get_variable("kernel_order_{}".format(order), shape=[dimension], dtype=tf.float64, initializer=tf.random_normal_initializer())
 
     def build_model(self):
         self._init_data_structs()
